delete saves and rename AI folder so mac doesn't throw silly errors because of case insensitivity
