--- src/AI/index.ts
@@ -1,2 +1,0 @@
-export * from './stages/startNewChat';
-export * from './stages/loadChat';
