+++ src/AIchat/index.ts
@@ -1,0 +1,3 @@
+export * from './commands/startNewChat';
+export * from './commands/loadChat';
+export * from './commands/deleteChat';
